api/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
python-dotenv
pandas
